Bug 1328337 - fix implicit conversion of glyph size from float to uint32_t in NativeFontResource::CreateScaledFont. r=bobowen
MozReview-Commit-ID: 9JcrfRYJ3ca

+++ gfx/2d/NativeFontResourceMac.cpp
@@ -50,7 +50,7 @@ NativeFontResourceMac::Create(uint8_t *aFontData, uint32_t aDataLength)
 }
 
 already_AddRefed<ScaledFont>
-NativeFontResourceMac::CreateScaledFont(uint32_t aIndex, uint32_t aGlyphSize)
+NativeFontResourceMac::CreateScaledFont(uint32_t aIndex, Float aGlyphSize)
 {
   RefPtr<ScaledFontBase> scaledFont = new ScaledFontMac(mFontRef, aGlyphSize);
